refactor: move window sizing logic
